#2391 Relocate a unit test
