+++ platform-pojo-bl/src/test/java/ua/com/fielden/platform/types/RichTextHtmlTest.java
@@ -3,8 +3,7 @@ package ua.com.fielden.platform.types;
 import org.junit.Test;
 
 import static org.assertj.core.api.Assertions.assertThat;
-import static org.junit.Assert.assertEquals;
-import static org.junit.Assert.assertThrows;
+import static org.junit.Assert.*;
 import static ua.com.fielden.platform.types.RichTextSanitiser.ERR_UNSAFE;
 
 public class RichTextHtmlTest {
@@ -709,6 +708,20 @@ After quote.
         assertSearchText("two three four", "<b>two </b> <i>three</i> four");
         assertSearchText("two three four", " <b>two </b> <i>three</i> four");
         assertSearchText("two three four", "<b> two </b> <i>three</i> four");
+    }
+
+    @Test
+    public void safe_HTML_that_becomes_unsafe_after_conversion_to_coreText_is_disallowed() {
+        final var html = "<code>&lt;img onerror=alert(1) /&gt;</code>";
+        final var htmlResult = RichTextSanitiser.sanitiseHtml(html);
+        assertTrue(htmlResult.isSuccessful());
+
+        final var invalidRichText = RichText.fromHtml(html);
+        assertThat(invalidRichText).isInstanceOf(RichText.Invalid.class);
+        assertThat(invalidRichText.isValid().getMessage()).isEqualTo("""
+                %s<extended/>Input contains unsafe HTML:
+                1. Tag [img] has violating attributes: onerror\
+                """.formatted(ERR_UNSAFE));
     }
 
     @Test
